refactor: Extracted XRContext and XRState into separate module
this way it'll be easier to mock a store provider without exposing XRContext to the outside
