+++ src/index.tsx
@@ -5,3 +5,4 @@ export * from './XR'
 export * from './XRController'
 export * from './XREvents'
 export * from './XRControllerModelFactory'
+export { type XRState } from './context'
